Fix log in ExecJava for SpooledDirectory Fix file size check (must be greater than 0)
