+++ src/main/java/org/waarp/openr66/protocol/utils/Version.java
@@ -3,7 +3,7 @@ package org.waarp.openr66.protocol.utils;
 /** Provides the version information of Waarp OpenR66. */
 public final class Version {
  /** The version identifier. */
- public static final String ID = "2.4.20";
+ public static final String ID = "2.4.21";
  /** Prints out the version identifier to stdout. */
  public static void main(String[] args) { System.out.println(ID); }
  private Version() { super(); }
